requirements.txt
streamlit
faiss-cpu
sentence-transformers
requests
python-docx
pandas
openpyxlc